Implement FUSE-based archive mounting
This allows you to mount an archive as a read-only filesystem to a specified directory by pressing the `l` key.

Currently, it tends to consume a lot of memory on large archives as a read call causes the entire file to be read into memory. There are some improvements that can be made here.